+++ src/main/java/edu/icet/repository/BirthdayPartyRepository.java
@@ -3,5 +3,5 @@ package edu.icet.repository;
 import edu.icet.entity.BirthdayPartyEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BirthdayPartyRepository extends JpaRepository<BirthdayPartyEntity, Integer> {
+public interface BirthdayPartyRepository extends JpaRepository<BirthdayPartyEntity, Long> {
 }
